refactor: split cloudflare publisher adapter
