+++ .dockerignore
@@ -11,7 +11,7 @@ docs
 examples
 output
 artifacthub-repo.yml
-CLAUDE.md
+AGENTS.md
 CODE_OF_CONDUCT.md
 CONTRIBUTING.md
 README.md
