+++ publisher/retry.go
@@ -1,0 +1,10 @@
+package publisher
+
+import (
+	"math"
+	"time"
+)
+
+func retryDelay(attempt int) time.Duration {
+	return time.Duration(math.Pow(2, float64(attempt))) * time.Second
+}
